Fix the kl_div docs
Fixes https://github.com/pytorch/pytorch/issues/57459

After discussing the linked issue, we resolved that `F.kl_div` computes
the right thing as to be consistent with the rest of the losses in
PyTorch.

To avoid any confusion, these docs add a note discussing how the PyTorch
implementation differs from the mathematical definition and the reasons
for doing so.

These docs also add an example that may further help understanding the
intended use of this loss.

ghstack-source-id: 6eb16cbbb6defdd3f700efd0283c277868ba5e8e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/67443