[inductor][do not land] attempt 1, fix addmm for scalar addend
addmm(scalar, tensor1, tensor2) fails, because the scalar causes some assertion failures. This is a first attempt to fix this.

This is bad, it will break for something like this:
```python
def fn(a, b)
    return torch.mm(a, b) + 3

a = torch.rand((8, 2))
b = torch.rand((2, 8))
torch._dynamo.optimize()(fn)(a, b)
```

other note is that addmm(inp, mat1, mat2) in eager pytorch does NOT accept a scalar for inp. So maybe the other way to try fixing this is to modify the pattern matching fx passes to not produce these types of addmm calls.