[WIP][profiler] Update kineto pin + silence kineto logging during _init_for_cuda_graphs()
A recent update to kineto exposes the logging API (kineto PR 757). This updates the kineto pin, so that we can use the logging API to disable logging while the _init_for_cuda_graphs is running.

Context: _init_for_cuda_graphs is pused just to pre-init CUPTI to work around cuda graphs + CUPTI issues, which is needed in certain versions of CUDA / CUDA drivers. In these cases we aren't actually profiling anything, so the logs about profiling are not useful.